+++ src/compliment/sources/classes.clj
@@ -1,20 +1,9 @@
-(ns compliment.sources.namespaces-and-classes
-  "Completion for namespace and class names."
+(ns compliment.sources.classes
+  "Completion for class names."
   (:require [compliment.sources :refer [defsource]]
             [compliment.utils :refer [fuzzy-matches?] :as utils]
-            [compliment.sources.class-members :refer [classname-doc]]))
-
-(defn nscl-symbol?
-  "Tests if prefix looks like a namespace or classname."
-  [^String x]
-  (and (re-matches #"[^\/\:]+" x)
-       (not (= (.charAt x 0) \.))))
-
-(defn nscl-matches?
-  "Tests if prefix partially matches a var name with periods as
-  separators."
-  [prefix namespace]
-  (fuzzy-matches? prefix namespace \.))
+            [compliment.sources.class-members :refer [classname-doc]]
+            [compliment.sources.namespaces :refer [nscl-symbol? nscl-matches?]]))
 
 ;;; Obtaining the list of classes
 
@@ -24,7 +13,7 @@
   (for [[_ ^Class val] (ns-map ns) :when (class? val)]
     (.getName val)))
 
-(defn all-classes-short-names
+(defn- all-classes-short-names
   "Returns a map where short classnames are matched with vectors with
   package-qualified classnames."
   []
@@ -71,19 +60,14 @@
              (all-classes-short-names)))
 
 (defn candidates
-  "Returns a list of namespace and classname completions."
+  "Returns a list of classname completions."
   [^String prefix, ns context]
   (when (nscl-symbol? prefix)
     (let [has-dot (> (.indexOf prefix ".") -1)
           import-ctx (analyze-import-context context)]
       (into []
             (comp cat (distinct))
-            [(for [ns-str (concat (map (comp name ns-name) (all-ns))
-                                  (map name (keys (ns-aliases ns))))
-                   :let [[literals prefix] (utils/split-by-leading-literals prefix)]
-                   :when (nscl-matches? prefix ns-str)]
-               {:candidate (str literals ns-str), :type :namespace})
-             (for [class-str (imported-classes ns)
+            [(for [class-str (imported-classes ns)
                    :when (nscl-matches? prefix class-str)]
                {:candidate class-str, :type :class})
              (cond (= import-ctx :root) (get-all-full-names prefix)
@@ -92,15 +76,6 @@
              (when (and (Character/isUpperCase (.charAt prefix 0))
                         (not import-ctx))
                (get-all-full-names prefix))
-             ;; If prefix doesn't contain a period, using fuziness produces too many
-             ;; irrelevant candidates.
-             (for [{:keys [^String ns-str, ^String file]} (utils/namespaces&files-on-classpath)
-                   :let [[literals prefix] (utils/split-by-leading-literals prefix)]
-                   :when (and (re-find #"\.cljc?$" file)
-                              (if has-dot
-                                (nscl-matches? prefix ns-str)
-                                (.startsWith ns-str prefix)))]
-               {:candidate (str literals ns-str), :type :namespace, :file file})
              ;; Fuzziness is too slow for all classes, so only startsWith. Also, if no
              ;; period in prefix, only complete root package names to maintain good
              ;; performance and not produce too many candidates.
@@ -115,18 +90,15 @@
                        :when (.startsWith root-pkg prefix)]
                    {:candidate (str root-pkg "."), :type :class})))]))))
 
-(defn doc [ns-or-class-str curr-ns]
-  (when (nscl-symbol? ns-or-class-str)
+(defn doc [class-str curr-ns]
+  (when (nscl-symbol? class-str)
     (let [strip-literals (comp second utils/split-by-leading-literals)
-          ns-or-class-sym (symbol (strip-literals ns-or-class-str))]
-      (if-let [ns (or (find-ns ns-or-class-sym)
-                      (get (ns-aliases curr-ns) ns-or-class-sym))]
-        (str ns "\n" (:doc (meta ns)) "\n")
-        (when-let [class (try (ns-resolve curr-ns ns-or-class-sym)
-                              (catch Exception ex nil))]
-          (when (= (type class) Class)
-            (classname-doc class)))))))
+          class-sym (symbol (second (utils/split-by-leading-literals class-str)))]
+      (when-let [class (try (ns-resolve curr-ns class-sym)
+                            (catch Exception ex nil))]
+        (when (class? class)
+          (classname-doc class))))))
 
-(defsource ::namespaces-and-classes
+(defsource ::classes
   :candidates #'candidates
   :doc #'doc)